Move row parsing into the table
