+++ readyset-vitess-data/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod field_parsing;
-pub mod row_parsing;
 pub mod vitess;
 
 pub use field_parsing::vstream_value_to_noria_value;
-pub use row_parsing::vstream_row_to_noria_row;
