+++ readyset-vitess-data/tests/row_parsing.rs
@@ -2,7 +2,6 @@ use std::sync::Arc;
 
 use readyset_data::DfValue;
 use readyset_vitess_data::vitess::Table;
-use readyset_vitess_data::vstream_row_to_noria_row;
 use vitess_grpc::query::Row;
 
 mod helpers;
@@ -16,16 +15,15 @@ fn products_table() -> Table {
 
 #[test]
 fn basics() {
-    let insert_row = Row {
+    let res = products_table().vstream_row_to_noria_row(&Row {
         lengths: vec![2, 10, 2],
         values: vec![
             112, 49, // p1
             112, 114, 111, 100, 117, 99, 116, 32, 35, 49, // product #1
             52, 50, // 42
         ],
-    };
+    });
 
-    let res = vstream_row_to_noria_row(&insert_row, &products_table());
     assert!(res.is_ok());
     assert_eq!(
         res.unwrap(),
@@ -39,12 +37,11 @@ fn basics() {
 
 #[test]
 fn null_fields() {
-    let insert_row = Row {
+    let res = products_table().vstream_row_to_noria_row(&Row {
         lengths: vec![3, -1, -1],
         values: vec![49, 48, 48], // 100
-    };
+    });
 
-    let res = vstream_row_to_noria_row(&insert_row, &products_table());
     assert!(res.is_ok());
 
     let res = res.unwrap();
